+++ polytope/datacube/datacube.py
@@ -47,3 +47,13 @@ class Datacube(ABC):
 
             xadatacube = XArrayDatacube(datacube, axis_options=axis_options)
             return xadatacube
+
+
+def configure_datacube_axis(options, name, values, datacube):
+    if options == {}:
+        DatacubeAxis.create_standard(name, values, datacube)
+    if "mapper" in options.keys():
+        from .datacube_mappers import DatacubeMapper
+        DatacubeMapper.create_mapper(options, name, datacube)
+    if "cyclic" in options.keys():
+        DatacubeAxis.create_cyclic(options, name, values, datacube)
